Refactor ImportEventsCommand and cleanup files
The code has been refactored to improve the efficiency of event importing, and a cleanup of redundant files and comments has been performed. Also, the 'hash' field has been added to the database table for events. The changes allow for improved tracking of changes and easier debugging with the new 'debug' option. Plus, a new 'removelock' option has also been introduced to remove lock files.