Refactor list to array destructuring and improve data check
The code has been updated to use array destructuring instead of list. This is done for queries and query constraints generated for preparing 'FindByEtKeysQuery' in the EventcontainerController. Furthermore, an additional check has been added to ensure 'uid' exists and is referenced only when it exists in the $data array, to avoid unexpected errors.